Version 0.2.0: use a HashMap to track borrows
This commit reimplements the ledger as a HashMap. An FxHashMap is used
to take advantage of the small size of the key.  The Mutex from the
standard library has been replaced with a ticketing spin lock from the
spin crate. The main reason for this change is that we don't want the
thread to be parked until it can acquire the mutex because that can
cause a deadlock to occur if another thread indicates that garbage must
be collected.

Benchmarks and a CI workflow file have also been added.